+++ strax/context.py
@@ -15,6 +15,8 @@ from collections import defaultdict
 from immutabledict import immutabledict
 from enum import IntEnum
 
+from strax import CutList
+
 
 export, __all__ = strax.exporter()
 __all__.extend(["RUN_DEFAULTS_KEY"])
@@ -189,7 +191,11 @@ class Context:
     _run_defaults_cache: dict
     storage: ty.List[strax.StorageFrontend]
 
-    def __init__(self, storage=None, config=None, register=None, register_all=None, **kwargs):
+    processors: ty.Mapping[str, strax.BaseProcessor]
+
+    def __init__(
+        self, storage=None, config=None, register=None, register_all=None, processors=None, **kwargs
+    ):
         """Create a strax context.
 
         :param storage: Storage front-ends to use. Can be:
@@ -200,7 +206,9 @@ class Context:
             applied to plugins
         :param register: plugin class or list of plugin classes to register
         :param register_all: module for which all plugin classes defined in it
-            will be registered.
+           will be registered.
+        :param processors: A mapping of processor names to classes to use for
+            data processing.
         Any additional kwargs are considered Context-specific options; see
         Context.takes_config.
 
@@ -224,12 +232,32 @@ class Context:
         if register is not None:
             self.register(register)
 
+        if processors is None:
+            processors = strax.PROCESSORS
+
+        if isinstance(processors, str):
+            processors = [processors]
+
+        if isinstance(processors, (list, tuple)):
+            ps = {}
+            for processor in processors:
+                if isinstance(processor, str) and processor in strax.PROCESSORS:
+                    ps[processor] = strax.PROCESSORS[processor]
+                elif isinstance(processor, strax.BaseProcessor):
+                    ps[processor.__name__] = processor
+                else:
+                    raise ValueError(f"Unknown processor {processor}")
+            processors = ps
+
+        self.processors = processors
+
     def new_context(
         self,
         storage=tuple(),
         config=None,
         register=None,
         register_all=None,
+        processors=None,
         replace=False,
         **kwargs,
     ):
@@ -253,7 +281,7 @@ class Context:
             config = strax.combine_configs(self.config, config, mode="update")
             kwargs = strax.combine_configs(self.context_config, kwargs, mode="update")
 
-        new_c = Context(storage=storage, config=config, **kwargs)
+        new_c = Context(storage=storage, config=config, processors=processors, **kwargs)
         if not replace:
             new_c._plugin_class_registry = self._plugin_class_registry.copy()
         new_c.register_all(register_all)
@@ -318,6 +346,9 @@ class Context:
             for x in plugin_class:
                 self.register(x)
             return
+
+        if not issubclass(plugin_class, strax.Plugin):
+            raise ValueError(f"Can only register subclasses of strax.Plugin, not {plugin_class}!")
 
         if not hasattr(plugin_class, "provides"):
             # No output name specified: construct one from the class name
@@ -424,6 +455,24 @@ class Context:
             if not len(plugins_to_deregister):
                 registry_changed = False
 
+    def get_data_kinds(self) -> ty.Tuple:
+        """Return two dictionaries:
+        1. one with all available data_kind as key and their data_types(list) as values
+        2. one with all available data_type as key and their data_kind(str) as values
+        """
+        data_kind_collection: ty.Dict[str, ty.List] = dict()
+        data_type_collection: ty.Dict[str, str] = dict()
+        for data_type in self._plugin_class_registry.keys():
+            plugin = self.__get_plugin("0", data_type)
+            if isinstance(plugin.data_kind, (dict, immutabledict)):
+                data_kind = plugin.data_kind[data_type]
+            else:
+                data_kind = plugin.data_kind
+            data_kind_collection.setdefault(data_kind, [])
+            data_kind_collection[data_kind].append(data_type)
+            data_type_collection[data_type] = data_kind
+        return data_kind_collection, data_type_collection
+
     def search_field(
         self,
         pattern: str,
@@ -501,7 +550,7 @@ class Context:
                                 # 'type'
                                 # >>> A().__class__.__name__
                                 # 'A'
-                                plug = plug()
+                                plug = plug()  # type: ignore
                             result += [f"{plug.__class__.__name__}.{attribute_name}"]
                             # Likely to be used several other times
                             break
@@ -575,6 +624,28 @@ class Context:
             if issubclass(x, strax.Plugin):
                 self.register(x)
 
+    def register_cut_list(self, cut_list):
+        """Register cut lists to strax context.
+
+        :param cut_list: cut lists to be registered. can be cutlist object or list/tuple of cutlist
+            objects
+
+        """
+        assert not isinstance(
+            cut_list, str
+        ), "Please don't put string... use cutlist object or list/tuple of cutlist objects"
+        if hasattr(cut_list, "__len__"):
+            for _cut_list in cut_list:
+                self.register_cut_list(_cut_list)
+        else:
+            for cut in cut_list.cuts:
+                # maybe cutlist within cutlist?
+                if CutList in cut.__bases__:
+                    self.register_cut_list(cut)
+                else:
+                    self.register(cut)
+            self.register(cut_list)
+
     def data_info(self, data_name: str) -> pd.DataFrame:
         """Return pandas DataFrame describing fields in data_name."""
         p = self._get_plugins((data_name,), run_id="0")[data_name]
@@ -599,14 +670,23 @@ class Context:
         plugin.setup()
         return plugin
 
+    @staticmethod
+    def hyperrun_id(run_id):
+        if run_id.startswith("__"):
+            _run_id = run_id[2:]
+        else:
+            _run_id = run_id
+        return _run_id
+
     def _set_plugin_config(self, p, run_id, tolerant=True):
         # Explicit type check, since if someone calls this with
         # plugin CLASSES, funny business might ensue
+        _run_id = self.hyperrun_id(run_id)
         assert isinstance(p, strax.Plugin)
         config = self.config.copy()
         for opt in p.takes_config.values():
             try:
-                opt.validate(config, run_id=run_id, run_defaults=self.run_defaults(run_id))
+                opt.validate(config, run_id=_run_id, run_defaults=self.run_defaults(_run_id))
             except strax.InvalidConfiguration:
                 if not tolerant:
                     raise
@@ -704,7 +784,7 @@ class Context:
                 continue
 
             requested_p = plugin.__copy__()
-            requested_p.run_id = run_id
+            requested_p.run_id = self.hyperrun_id(run_id)
 
             # Re-use only one instance if the plugin is multi output
             for provides in strax.to_str_tuple(requested_p.provides):
@@ -774,7 +854,7 @@ class Context:
 
         plugin = self._plugin_class_registry[data_type]()
 
-        plugin.run_id = run_id
+        plugin.run_id = self.hyperrun_id(run_id)
 
         # The plugin may not get all the required options here
         # but we don't know if we need the plugin yet
@@ -816,8 +896,6 @@ class Context:
             # drop the parents config from the lineage
             configs = {}
 
-            # Getting information about the parent:
-            parent_class = plugin.__class__.__bases__[0]
             # Get all parent options which are overwritten by a child:
             parent_options = [
                 option.parent_option_name
@@ -837,7 +915,8 @@ class Context:
                     configs[option_name] = v
 
             # Also adding name and version of the parent to the lineage:
-            configs[parent_class.__name__] = parent_class.__version__
+            for parent_class in plugin.__class__.__bases__:
+                configs[parent_class.__name__] = parent_class.__version__
 
             plugin.lineage = {
                 last_provide: (plugin.__class__.__name__, plugin.version(run_id), configs)
@@ -906,6 +985,10 @@ class Context:
         """
         return sorted(self.storage, key=lambda x: x.storage_type)
 
+    def writable_storage(self) -> ty.List[strax.StorageFrontend]:
+        """Return list of writable storage frontends."""
+        return [s for s in self.storage if not s.readonly]
+
     def _get_partial_loader_for(self, key, time_range=None, chunk_number=None):
         """Get partial loaders to allow loading data later.
 
@@ -941,6 +1024,7 @@ class Context:
         save=tuple(),
         time_range=None,
         chunk_number=None,
+        multi_run_progress_bar=False,
     ) -> strax.ProcessorComponents:
         """Return components for setting up a processor.
 
@@ -955,6 +1039,10 @@ class Context:
                 raise ValueError(f"Plugin names must be more than one letter, not {t}")
 
         plugins = self._get_plugins(targets, run_id)
+
+        allow_hyperruns = [plugins[target_i].allow_hyperrun for target_i in targets]
+        if sum(allow_hyperruns) != 0 and sum(allow_hyperruns) != len(targets):
+            raise ValueError("Cannot mix plugins that allow hyperruns with those that do not.")
 
         # Get savers/loaders, and meanwhile filter out plugins that do not
         # have to do computation. (their instances will stick around
@@ -981,17 +1069,26 @@ class Context:
                 key, chunk_number=chunk_number, time_range=time_range
             )
 
-            _is_superrun = run_id.startswith("_") and not target_plugin.provides[0].startswith(
-                "_temp"
-            )
-            if not loader and _is_superrun:
+            _is_temp = not target_plugin.provides[0].startswith("_temp")
+            _is_superrun = run_id.startswith("_") and _is_temp
+            _is_hyperrun = run_id.startswith("__") and _is_temp
+            # hyperrun is always superrun
+            allow_hyperrun = plugins[target_i].allow_hyperrun
+            if not loader and (
+                (_is_superrun and not _is_hyperrun) or (_is_hyperrun and not allow_hyperrun)
+            ):
                 if time_range is not None:
                     raise NotImplementedError("time range loading not yet supported for superruns")
 
                 sub_run_spec = self.run_metadata(run_id, "sub_run_spec")["sub_run_spec"]
 
                 # Make subruns if they do not exist.
-                self.make(list(sub_run_spec.keys()), target_i, save=(target_i,))
+                self.make(
+                    list(sub_run_spec.keys()),
+                    target_i,
+                    save=(target_i,),
+                    multi_run_progress_bar=multi_run_progress_bar,
+                )
 
                 ldrs = []
                 for subrun in sub_run_spec:
@@ -1084,12 +1181,17 @@ class Context:
             current_plugin_to_savers = [target_i]
             if (
                 not self._target_should_be_saved(
-                    target_plugin, target_i, targets, save, loader, _is_superrun
+                    target_plugin,
+                    target_i,
+                    targets,
+                    save,
+                    loader,
+                    _is_superrun and not _is_hyperrun,
                 )
                 and not self.context_config["storage_converter"]
             ):
                 if len(target_plugin.provides) > 1:
-                    # In case the plugin has more then a single provides we also have to check
+                    # In case the plugin has more than a single provides we also have to check
                     # whether any of the other data_types should be stored. Hence only remove
                     # the current traget from the list of plugins_to_savers.
                     current_plugin_to_savers = []
@@ -1116,13 +1218,16 @@ class Context:
                 return
 
             # Save the target and any other outputs of the plugin.
-            if _is_superrun:
+            if _is_superrun and not _is_hyperrun:
                 # In case of a superrun we are only interested in the specified targets
                 # and not any other stuff provided by the corresponding plugin.
                 savers = self._add_saver(
-                    savers, target_i, target_plugin, _is_superrun, loading_this_data
+                    savers, target_i, run_id, target_plugin, _is_superrun, loading_this_data
                 )
-            else:
+            elif not _is_hyperrun or allow_hyperrun:
+                # only save if we are not in a hyperrun or the plugin allows hyperruns
+                # otherwise we will see error at Chunk.concatenate
+                # but anyway the data is should already been made
                 for d_to_save in set(current_plugin_to_savers + list(target_plugin.provides)):
                     key = self.key_for(run_id, d_to_save)
                     loader = self._get_partial_loader_for(
@@ -1131,7 +1236,12 @@ class Context:
 
                     if (
                         not self._target_should_be_saved(
-                            target_plugin, d_to_save, targets, save, loader, _is_superrun
+                            target_plugin,
+                            d_to_save,
+                            targets,
+                            save,
+                            loader,
+                            _is_superrun and not _is_hyperrun,
                         )
                         and not self.context_config["storage_converter"]
                     ) or savers.get(d_to_save):
@@ -1141,7 +1251,7 @@ class Context:
                         continue
 
                     savers = self._add_saver(
-                        savers, d_to_save, target_plugin, _is_superrun, loading_this_data
+                        savers, d_to_save, run_id, target_plugin, _is_superrun, loading_this_data
                     )
 
         for target_i in targets:
@@ -1178,6 +1288,7 @@ class Context:
         self,
         savers: dict,
         d_to_save: str,
+        run_id,
         target_plugin: strax.Plugin,
         _is_superrun: bool,
         loading_this_data: bool,
@@ -1194,7 +1305,7 @@ class Context:
         :return: Updated savers dictionary.
 
         """
-        key = strax.DataKey(target_plugin.run_id, d_to_save, target_plugin.lineage)
+        key = strax.DataKey(run_id, d_to_save, target_plugin.lineage)
         for sf in self._sorted_storage:
             if sf.readonly:
                 continue
@@ -1218,7 +1329,7 @@ class Context:
             try:
                 saver = sf.saver(
                     key,
-                    metadata=target_plugin.metadata(target_plugin.run_id, d_to_save),
+                    metadata=target_plugin.metadata(run_id, d_to_save),
                     saver_timeout=self.context_config["saver_timeout"],
                 )
                 # Now that we are surely saving, make an entry in savers
@@ -1353,7 +1464,7 @@ class Context:
     def get_iter(
         self,
         run_id: str,
-        targets: ty.Union[ty.Tuple[str], ty.List[str]],
+        targets,
         save=tuple(),
         max_workers=None,
         time_range=None,
@@ -1366,7 +1477,9 @@ class Context:
         drop_columns=None,
         allow_multiple=False,
         progress_bar=True,
+        multi_run_progress_bar=True,
         _chunk_number=None,
+        processor=None,
         **kwargs,
     ) -> ty.Iterator[strax.Chunk]:
         """Compute target for run_id and iterate over results.
@@ -1421,7 +1534,12 @@ class Context:
                 raise RuntimeError(f"Cannot allow_multiple in lazy mode or with long timeouts.")
 
         components = self.get_components(
-            run_id, targets=targets, save=save, time_range=time_range, chunk_number=_chunk_number
+            run_id,
+            targets=targets,
+            save=save,
+            time_range=time_range,
+            chunk_number=_chunk_number,
+            multi_run_progress_bar=multi_run_progress_bar,
         )
 
         # Cleanup the temp plugins
@@ -1429,8 +1547,17 @@ class Context:
             if k.startswith("_temp"):
                 del self._plugin_class_registry[k]
 
+        if processor is None:
+            processor = list(self.processors)[0]
+
+        if isinstance(processor, str):
+            processor = self.processors[processor]
+
+        if not hasattr(processor, "iter"):
+            raise ValueError("Processors must implement a iter methed.")
+
         seen_a_chunk = False
-        generator = strax.ThreadedMailboxProcessor(
+        generator = processor(
             components,
             max_workers=max_workers,
             allow_shm=self.context_config["allow_shm"],
@@ -1545,7 +1672,7 @@ class Context:
         pbar.mbs.append((nbytes / 1e6) / seconds_per_chunk)
         mbs = np.mean(pbar.mbs)
         if mbs < 1:
-            rate = f"{mbs*1000:.1f} kB/s"
+            rate = f"{mbs * 1000:.1f} kB/s"
         else:
             rate = f"{mbs:.1f} MB/s"
         postfix = f"#{n_chunks} ({seconds_per_chunk:.2f} s). {rate}"
@@ -1584,7 +1711,7 @@ class Context:
                 **kwargs,
             )
 
-        if _skip_if_built and self.is_stored(run_id, targets):
+        if _skip_if_built and self.is_stored(run_ids[0], targets):
             return
 
         for _ in self.get_iter(run_ids[0], targets, save=save, max_workers=max_workers, **kwargs):
@@ -2374,6 +2501,57 @@ class Context:
             for data_type, _hash, save_when, version in hashes
         }
 
+    @property
+    def root_data_types(self):
+        """Root data_type that does not depend on anything."""
+        _root_data_types = set()
+        for k, v in self._plugin_class_registry.items():
+            if not v.depends_on:
+                _root_data_types |= set((k,))
+        return _root_data_types
+
+    @property
+    def inverse_tree(self):
+        """Inverse tree whose key is depends_on and value is provides."""
+        _inverse_tree = dict()
+        for k, v in self._plugin_class_registry.items():
+            for d in v().depends_on:
+                _inverse_tree.setdefault(d, [])
+                _inverse_tree[d] += v().provides
+        return _inverse_tree
+
+    def check_hyperrun(self):
+        """Raise if non-hyperrun plugins depends on hyperrun plugins."""
+        inverse_tree = self.inverse_tree
+
+        # define a recursive function to check if all the dependencies support hyperruns
+        def check_support_hyperrun(data_type, checked=set(), seen_allow=None):
+            if data_type in checked:
+                return checked
+            if self._plugin_class_registry[data_type].allow_hyperrun:
+                seen_allow = data_type
+            if seen_allow and not self._plugin_class_registry[data_type].allow_hyperrun:
+                raise ValueError(
+                    f"Already support hyperruns in {seen_allow}, "
+                    f"but it's dependency {data_type} does not support hyperruns."
+                )
+            checked |= set((data_type,))
+            if data_type not in inverse_tree:
+                return checked
+            for d in inverse_tree[data_type]:
+                checked |= check_support_hyperrun(d, checked, seen_allow)
+            return checked
+
+        # use checked set to record the data_type that has been checked
+        # to shorten the time of checking
+        checked = set()
+        for data_type in self.root_data_types:
+            if self._plugin_class_registry[data_type].allow_hyperrun:
+                seen_allow = data_type
+            else:
+                seen_allow = None
+            checked |= check_support_hyperrun(data_type, checked, seen_allow)
+
     @classmethod
     def add_method(cls, f):
         """Add f as a new Context method."""
@@ -2404,10 +2582,11 @@ select_docs = """
 :param multi_run_progress_bar: Display a progress bar for loading multiple runs
 """
 
-get_docs = (
-    """
+get_docs = """
 :param run_id: run id to get
 :param targets: list/tuple of strings of data type names to get
+:param ignore_errors: Return the data for the runs that successfully loaded, even if some runs
+        failed executing.
 :param save: extra data types you would like to save
     to cache, if they occur in intermediate computations.
     Many plugins save automatically anyway.
@@ -2423,9 +2602,10 @@ get_docs = (
 :param run_id_as_bytes: Boolean if true uses byte string instead of an
     unicode string added to a multi-run array. This can save a lot of
     memory when loading many runs.
+:param processor: Name of the processor to use. If not specified, the
+    first processor from the context's processor list is used.
 """
-    + select_docs
-)
+get_docs += select_docs
 
 for attr in dir(Context):
     attr_val = getattr(Context, attr)